+++ src/utils/device-utils.ts
@@ -1,4 +1,4 @@
-// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2020-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 import { DeviceType } from '../types';
@@ -46,3 +46,8 @@ export const isOptionActive = (
   }
   return currentDeviceId === meetingManagerDeviceId;
 };
+
+// TODO: Remove this and use DefaultBrowserBehavior.supportsSetSinkId from JS SDK v2.x
+export const supportsSetSinkId = (): boolean => {
+  return 'setSinkId' in HTMLAudioElement.prototype;
+}