+++ CHANGELOG.md
@@ -8,8 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Fixed
+- Make components compatible with JS SDK 2.x
+
+### Added
+
+### Changed
+
+### Removed
+
+## [1.6.0] - 2020-12-14
+
+### Fixed
 - Remove needless camera selection when joining meeting
 - [Demo] Fix demo test speakers
+- Fix meeting manager to handle `setSinkId` error
 
 ### Added
 - [Demo] Add Chat Demo app